// src/models/stnk.ts
export interface Stnk {
    id?: string;
    vehicleId?: string;
    licensePlate?: string;
    brand?: string;
    model?: string;
    type?: string;
    bodyModel?: string;
    manufactureYear?: string;
    engineCapacity?: string;
    chassisNumber?: string;
    engineNumber?: string;
    stnkNumber?: string;
    color?: string;
    fuelType?: string;
    tnkbColor?: string;
    registrationYear?: string;
    bpkbNumber?: string;
    registrationOrderNumber?: string;
    locationCode?: string;
    validUntil?: string;
}